Rename from CreateUser to RegisterUser
